requirements.txt
torch
torchvision
opencv-python
streamlit
roboflow
pycocotools
matplotlib
pillow
tqdm
requests